chore: 优化脚本,不在编译字库,进一步减小包体积

+++ build.sh
@@ -6,33 +6,37 @@ set -e
 rime_version=1.8.5
 rime_git_hash=08dd95f
 
-rime_archive="rime-${rime_git_hash}-macOS.tar.bz2"
-rime_download_url="https://github.com/rime/librime/releases/download/${rime_version}/${rime_archive}"
+if [[ ! -d .deps ]] 
+then
+  rime_archive="rime-${rime_git_hash}-macOS.tar.bz2"
+  rime_download_url="https://github.com/rime/librime/releases/download/${rime_version}/${rime_archive}"
 
-rime_deps_archive="rime-deps-${rime_git_hash}-macOS.tar.bz2"
-rime_deps_download_url="https://github.com/rime/librime/releases/download/${rime_version}/${rime_deps_archive}"
+  rime_deps_archive="rime-deps-${rime_git_hash}-macOS.tar.bz2"
+  rime_deps_download_url="https://github.com/rime/librime/releases/download/${rime_version}/${rime_deps_archive}"
 
-rm -rf .deps && mkdir -p .deps && (
-    cd .deps
-    [ -z "${no_download}" ] && curl -LO "${rime_download_url}"
-    tar --bzip2 -xf "${rime_archive}"
-    [ -z "${no_download}" ] && curl -LO "${rime_deps_download_url}"
-    tar --bzip2 -xf "${rime_deps_archive}"
-)
+  rm -rf .deps && mkdir -p .deps && (
+      cd .deps
+      [ -z "${no_download}" ] && curl -LO "${rime_download_url}"
+      tar --bzip2 -xf "${rime_archive}"
+      [ -z "${no_download}" ] && curl -LO "${rime_deps_download_url}"
+      tar --bzip2 -xf "${rime_deps_archive}"
+  )
+fi
 
-# 输入方案临时目录
+# # 输入方案临时目录
 OUTPUT=".tmp"
 DST_PATH="$OUTPUT/SharedSupport"
 rm -rf .plum $OUTPUT
 mkdir -p $DST_PATH
 mkdir -p $DST_PATH/lua
 mkdir -p $DST_PATH/opencc
+mkdir -p $DST_PATH/build
 git clone --depth 1 https://github.com/rime/plum.git $OUTPUT/.plum
 
 # 可以在这里添加rime的开源输入方案
 # https://github.com/rime/rime-double-pinyin.git
 # for package in essay prelude rime-double-pinyin; do
-for package in essay prelude; do
+for package in prelude; do
   bash $OUTPUT/.plum/scripts/install-packages.sh "${package}" $DST_PATH
 done
 
@@ -45,7 +49,10 @@ rm -rf $OUTPUT/.clover && mkdir -p $OUTPUT/.clover && (
     [ -z "${no_download}" ] && curl -LO "${rime_cloverpinyin_download_url}"
     unzip "${rime_cloverpinyin_archive}" -d .
     rm -rf ${rime_cloverpinyin_archive}
-) && cp -R $OUTPUT/.clover/* $DST_PATH/
+) && \
+cp -R $OUTPUT/.clover/*.yaml $DST_PATH/ && \
+cp -R $OUTPUT/.clover/build/* $DST_PATH/build/ && \
+cp -R $OUTPUT/.clover/opencc/* ${DST_PATH}/opencc/
 
 # 内置极点五笔, qq五笔, 小鹤双拼
 #internalSchemas=("wubi86_jidian" "wubi86_qq" "double_pinyin")
@@ -58,9 +65,9 @@ done
 # 五笔方案
 # 方案来源: https://github.com/networm/Rime
 rm -rf $OUTPUT/.networm && \
-  git clone https://github.com/networm/Rime $OUTPUT/.networm && (
+  git clone --depth 1 https://github.com/networm/Rime $OUTPUT/.networm && (
     cd $OUTPUT/.networm
-    rm -rf .git .gitignore README.md rime.lua default.custom.yaml
+    rm -rf .git .gitignore README.md rime.lua default.custom.yaml weasel.custom.yaml
   ) && cp -R $OUTPUT/.networm/* $DST_PATH/
 
 # 星空键道
@@ -75,28 +82,24 @@ rm -rf $OUTPUT/.rime_jd && \
 # 星猫键道
 # 方案来源: https://github.com/wzxmer/xkjd6-rime 
 rm -rf $OUTPUT/.xmjd6 && \
-  git clone https://github.com/wzxmer/xkjd6-rime $OUTPUT/.xmjd6 && (
-    cd $OUTPUT/.xmjd6
-    rm -rf README.md weasel.custom.yaml default.custom.yaml
-  ) && 
-  cp $OUTPUT/.xmjd6/xmjd6.*.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/xmjd_W.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/xmjd_Y.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/xmjd_Z.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/xmjd6_user.txt ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/xmjd6cx.*.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/xmjd6dz.*.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/liangfen.*.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/pinyin_simp.*.yaml ${DST_PATH} && \
-  cp $OUTPUT/.xmjd6/lua/* ${DST_PATH}/lua/ && \
-  cp $OUTPUT/.xmjd6/opencc/* ${DST_PATH}/opencc/ && \
-  cat $OUTPUT/.xmjd6/rime.lua >> ${DST_PATH}/rime.lua
+  git clone --depth 1 https://github.com/wzxmer/xkjd6-rime $OUTPUT/.xmjd6 && (
+    cd $OUTPUT/.xmjd6/xmjd6
+    rm -rf README.md default.custom.yaml weasel.custom.yaml
+  ) && \
+  cp $OUTPUT/.xmjd6/xmjd6/*.yaml ${DST_PATH} && \
+  cp $OUTPUT/.xmjd6/xmjd6/*.txt ${DST_PATH} && \
+  cp $OUTPUT/.xmjd6/xmjd6/lua/*.lua ${DST_PATH}/lua/ && \
+  cp $OUTPUT/.xmjd6/xmjd6/opencc/* ${DST_PATH}/opencc/ && \
+  cat $OUTPUT/.xmjd6/xmjd6/rime.lua >> ${DST_PATH}/rime.lua
 
 # 宇浩输入法
 # 方案来源：https://github.com/forFudan/yuhao
 rm -rf $OUTPUT/.yuhao && \
-  git clone https://github.com/forFudan/yuhao $OUTPUT/.yuhao && \
-  cp $OUTPUT/.yuhao/schema/yuhao*.yaml ${DST_PATH} && \
+  git clone --depth 1 https://github.com/forFudan/yuhao $OUTPUT/.yuhao && (
+    cd $OUTPUT/.yuhao/schema
+    rm -rf default.custom.yaml
+  ) &&  \
+  cp $OUTPUT/.yuhao/schema/*.yaml ${DST_PATH} && \
   cp -r $OUTPUT/.yuhao/schema/lua/* ${DST_PATH}/lua/ && \
   cp $OUTPUT/.yuhao/schema/opencc/* ${DST_PATH}/opencc/ && \
   echo '-- 宇浩输入法' >> ${DST_PATH}/rime.lua && \
@@ -111,8 +114,8 @@ rm -rf $OUTPUT/.yuhao && \
 pushd "${DST_PATH}" > /dev/null
 
 # 减少essay.txt词库
-awk '($2 >= 500) {print}' essay.txt > essay.txt.min
-mv essay.txt.min essay.txt
+# awk '($2 >= 500) {print}' essay.txt > essay.txt.min
+# mv essay.txt.min essay.txt
 
 # sed -n '{
 #   s/^version: \(["]*\)\([0-9.]*\)\(["]*\)$/version: \1\2.minimal\3/
@@ -130,7 +133,9 @@ for schema in *.schema.yaml; do
 done
 
 # 隐藏五笔方案依赖的pinyin_simp.schema.yaml
-ls *.schema.yaml| grep -v pinyin_simp.schema.yaml | sed 's/^\(.*\)\.schema\.yaml/  - schema: \1/' > schema_list.yaml
+ls *.schema.yaml | grep -v pinyin_simp.schema.yaml | grep -v liangfen.schema.yaml \
+  | grep -v liangfen.schema.yaml | grep -v xmjddz.schema.yaml \
+  | sed 's/^\(.*\)\.schema\.yaml/  - schema: \1/' > schema_list.yaml
 # 这里不需要只替换luna_pinyin
 # grep -Ff schema_list.yaml default.yaml > schema_list.yaml.min
 # mv schema_list.yaml.min schema_list.yaml
@@ -149,10 +154,11 @@ popd > /dev/null
 # 先将squirrel.yaml拷贝出来
 cp -R SharedSupport/* ${DST_PATH}/ 
 
-# 一定要提前编译
-export DYLD_FALLBACK_LIBRARY_PATH=$DYLD_FALLBACK_LIBRARY_PATH:$PWD/.deps/dist/lib
+# TODO: 提前编译
+# export DYLD_FALLBACK_LIBRARY_PATH=$DYLD_FALLBACK_LIBRARY_PATH:$PWD/.deps/dist/lib
+# .deps/dist/bin/rime_deployer --build ${DST_PATH}
+
 cp -R .deps/share/opencc ${DST_PATH}
-.deps/dist/bin/rime_deployer --build ${DST_PATH}
 
 # 压缩输入方案
 (
